ECA-6904 add tests
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@28994 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/ejbca-rest-api/src/org/ejbca/ui/web/rest/api/resources/CertificateResource.java
@@ -310,7 +310,7 @@ public class CertificateResource extends BaseRestResource {
     @Path("/expire")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getCertificatesAboutToExpire(@Context HttpServletRequest requestContext,
-                                                 @QueryParam("days") int days,
+                                                 @QueryParam("days") long days,
                                                  @QueryParam("offset") int offset,
                                                  @QueryParam("maxNumberOfResults") int maxNumberOfResults) {
         if (requestContext == null) {
